test: split compose service coverage part
